requirements.txt
transformer-lens==1.10.0
transformers==4.35.2
jupyter==1.0.0
plotly==5.18.0
plotly-express==0.4.1
nbformat==5.9.2
ipykernel==6.27.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
pylint==3.0.2
black==23.11.0
pytest==7.4.3
torchvision==0.17.1
timm==0.9.16
datasets==2.17.1
streamlit==1.31.1
diffusers==0.26.3
leap-ie == 0.2.4
git+https://github.com/callummcdougall/eindex.git
seaborn
umap-learn